fix: no ``.destroy()`` was implemented and could cause ckeditor exceptions on ``do_action()`` of openerp. (fixes #5)
CKEditor have probably some stuff (at least some timeOut functions) that needs
to be destroyed cleanly when reloading a widget. ``.do_action()`` will trigger
the ``.destroy()`` method of the widget when reloading. I added an additional
verification to ``.is_syntax_valid()`` which could also be related to reload
issues of the widget.